Moved some protocol data to tubespeek.  Partial code for sharing files
This code does not share correctly, but it still shares file lists
as well as sets up some connections.  Have not been able to get the
actual file transfer to work yet, but its getting there.

Also updated the file so that it tries to build xob file from data
to be shared.  Causes a crash on any journal object hat has the
get_file_path() return empty.