Converted internal ID's to sha1.  Added duplicate check.
Now all entires have a sha1 id.  This id will be either the activity
id that is stored in the metadata if it exists.  Otherwise it will
be the sha1 hash of the file it is trying to share.  If there is no file
it just generates a random hash based on the time as I have no idea what
kind of file it would be.