Using instance dir instead of tmp.
I just learned that the tmp directory in our activity is held in
ram and the instance directory is the same as the tmp directory but
held in flash memory.  Because of this, I figured it would be wise
to store the files in the instance directory instead of the tmp.