Basic permissions: download only, add/remove, admin.
The first user to connect to the server is added as an admin.  Admins
will get a "Server Settings" button on their UI.  When clicked, they
will be presented with a list of all users ever connected to the server.
There they can give them upload/remove and admin privileges if desired.